+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Write your changes below this section. `npm version [major|minor|patch]` will au
 ### added
 
 - coarse level in demo
+- flexible vector field density
 
 ## [1.0.5] - 2021-05-26